+++ src/models/Duration.kt
@@ -2,6 +2,6 @@ package models
 
 import helpers.toTimestamp
 
-class Duration(val startTime: Int, var endTime: Int){
+class Duration(val startTime: Long, var endTime: Long){
     override fun toString(): String = "D:(${startTime.toTimestamp()}, ${endTime.toTimestamp()})"
 }
